add tomcrypt_private.h
